+++ core/main.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+#include "operomnia-1/operomnia.h"
+#include "operomnia-1/vectors.h"
+
+int main ( int argc, char ** arg ) {
+  return 0;
+}
