+++ core/operomnia-1/mouse.h
@@ -1,0 +1,23 @@
+#ifndef H_MOUSE
+#define H_MOUSE
+
+#include <stdio.h>
+#include "vectors.h"
+
+typedef struct _mouse_buttons {
+  bool left_button;
+  bool right_button;
+  bool middle_button;
+} mouse_buttons;
+
+typedef struct _helper_data helper_data;
+
+static mouse_buttons empty_buttons;
+
+mouse_buttons get_mouse_buttons( helper_data * in_data );
+
+mouse_buttons get_mouse_pressed( helper_data * in_data );
+
+vector get_mouse_pos( helper_data * in_data );
+
+#endif
